Add localforace, better separation for redux
