+++ frontend/package.json
@@ -41,6 +41,7 @@
     "html-webpack-plugin": "2.29.0",
     "identity-obj-proxy": "^3.0.0",
     "jest": "^21.0.1",
+    "localforage": "^1.5.0",
     "node-sass": "^4.5.3",
     "object-assign": "4.1.1",
     "postcss-flexbugs-fixes": "3.2.0",
