+++ frontend/src/common/initial-state/auth.js
@@ -1,0 +1,7 @@
+const initialAuthState = {
+    authenticated: false,
+    email: undefined,
+    refresh: undefined,
+    access: undefined,
+};
+export default initialAuthState;
